+++ NEWS.rst
@@ -6,6 +6,10 @@ every change, see the Git log.
 
 Latest
 ------
+* Minor: Added example and documentation for network.
+* Minor: Added network alias methods.
+* Minor: Added basic example for stream writer and reader.
+* Minor: Added << operator for stream writer and >> for stream reader.
 * Minor: Updated waf.
 
 12.0.0