+++ NEWS.rst
@@ -7,6 +7,10 @@ every change, see the Git log.
 Latest
 ------
 * Minor: Improved api of the detail stream class.
+* Minor: Added example and documentation for network.
+* Minor: Added network alias methods.
+* Minor: Added basic example for stream writer and reader.
+* Minor: Added << operator for stream writer and >> for stream reader.
 * Minor: Updated waf.
 
 12.0.0