+++ NEWS.rst
@@ -6,6 +6,10 @@ every change, see the Git log.
 
 Latest
 ------
+* tdb
+
+12.1.0
+------
 * Minor: Improved api of the detail stream class.
 * Minor: Added example and documentation for network.
 * Minor: Added network alias methods.